minor challenges/AAS/requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
pyparsing==3.1.2
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
